# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Yandex Smart Home proxy API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

/**
 * @nullable
 */
export type DevicePropertyParameters = { [key: string]: unknown } | null;

/**
 * @nullable
 */
export type DevicePropertyState = { [key: string]: unknown } | null;

export interface DeviceProperty {
  type: string;
  retrievable?: boolean;
  reportable?: boolean;
  /** @nullable */
  parameters?: DevicePropertyParameters;
  /** @nullable */
  state?: DevicePropertyState;
}

/**
 * @nullable
 */
export type DeviceCapabilityParameters = { [key: string]: unknown } | null;

/**
 * @nullable
 */
export type DeviceCapabilityState = { [key: string]: unknown } | null;

export interface DeviceCapability {
  type: string;
  retrievable?: boolean;
  reportable?: boolean;
  /** @nullable */
  parameters?: DeviceCapabilityParameters;
  /** @nullable */
  state?: DeviceCapabilityState;
}

export type DeviceQuasarInfo = { [key: string]: unknown };

export interface Device {
  id: string;
  name: string;
  aliases?: string[];
  type: string;
  external_id?: string;
  skill_id?: string;
  /** @nullable */
  room_id?: string | null;
  groups?: string[];
  capabilities?: DeviceCapability[];
  properties?: DeviceProperty[];
  /** @nullable */
  household_id?: string | null;
  quasar_info?: DeviceQuasarInfo;
}

export interface Room {
  id: string;
  name: string;
  household_id: string;
  devices?: string[];
}

export interface Household {
  id: string;
  name: string;
}

export type YandexUserInfoGroupsItem = { [key: string]: unknown };

export type YandexUserInfoScenariosItem = { [key: string]: unknown };

export interface YandexUserInfo {
  status: string;
  request_id?: string;
  rooms?: Room[];
  groups?: YandexUserInfoGroupsItem[];
  devices?: Device[];
  scenarios?: YandexUserInfoScenariosItem[];
  households?: Household[];
}

export type DeviceActionCapabilityState = { [key: string]: unknown };

export interface DeviceActionCapability {
  type: string;
  state: DeviceActionCapabilityState;
}

export interface DeviceActionInput {
  actions: DeviceActionCapability[];
}

export type DeviceActionResultDevicesItem = { [key: string]: unknown };

export interface DeviceActionResult {
  status: string;
  request_id?: string;
  devices?: DeviceActionResultDevicesItem[];
}

export type ScenarioTriggersItem = { [key: string]: unknown };

export type ScenarioStepsItem = { [key: string]: unknown };

export interface Scenario {
  id: string;
  name: string;
  is_active?: boolean;
  triggers?: ScenarioTriggersItem[];
  steps?: ScenarioStepsItem[];
}

export interface ScenariosResponse {
  status: string;
  request_id?: string;
  scenarios?: Scenario[];
}

export interface ScenarioActionResult {
  status: string;
  request_id?: string;
}

export type WeatherDataCurrent = {
  temperature_2m: number;
  relative_humidity_2m: number;
  weather_code: number;
  wind_speed_10m?: number;
  /** @nullable */
  precipitation_probability?: number | null;
};

export type WeatherDataHourly = {
  precipitation_probability?: number[];
  time?: string[];
};

export interface WeatherData {
  current: WeatherDataCurrent;
  hourly?: WeatherDataHourly;
  /** @nullable */
  city?: string | null;
  latitude: number;
  longitude: number;
}

export type GetWeatherParams = {
latitude?: number;
longitude?: number;
};