# lib/api-zod/src/generated/types/deviceCapability.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Yandex Smart Home proxy API
 * OpenAPI spec version: 0.1.0
 */
import type { DeviceCapabilityParameters } from './deviceCapabilityParameters';
import type { DeviceCapabilityState } from './deviceCapabilityState';

export interface DeviceCapability {
  type: string;
  retrievable?: boolean;
  reportable?: boolean;
  /** @nullable */
  parameters?: DeviceCapabilityParameters;
  /** @nullable */
  state?: DeviceCapabilityState;
}