# lib/api-zod/src/generated/types/deviceProperty.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Yandex Smart Home proxy API
 * OpenAPI spec version: 0.1.0
 */
import type { DevicePropertyParameters } from './devicePropertyParameters';
import type { DevicePropertyState } from './devicePropertyState';

export interface DeviceProperty {
  type: string;
  retrievable?: boolean;
  reportable?: boolean;
  /** @nullable */
  parameters?: DevicePropertyParameters;
  /** @nullable */
  state?: DevicePropertyState;
}